Extract serializers to separate file
